decode::careful: Bring this closer to completion

+++ src/decode/mod.rs
@@ -2,7 +2,7 @@ use instruction;
 
 pub mod careful;
 
-pub type DecodeResult<E> = Result<(instruction::Instruction, u8), E>;
+pub type DecodeResult<E> = Result<(instruction::Instruction, usize), E>;
 
 pub trait Decoder<E> {
     fn decode(&self, pc: &[u16]) -> DecodeResult<E>;
